refactor: functions for query
