+++ src/db/professionalQuerys.js
@@ -1,0 +1,22 @@
+const { Profissionais, Servicos } = require('../models');
+const treatData = require('./treatData')
+
+const listAll = () => Profissionais.findAll()
+
+let getProfissional = (paramId) => {
+    return Profissionais.findOne({
+        where: {
+            id: paramId
+        }, include: [{
+            model: Servicos,
+            attributes: ['nome']
+
+        }]
+    })
+}
+
+
+module.exports = {
+    listAll,
+    getProfissional,
+}
